+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # Changelog **TODOvue Search**
+
+## v1.0.0
+### Dependencies Update
+* Update `todovue-css` to version `1.5.0`
+* Update `todovue-demo` to version `2.2.1`
+* Update `todovue-search` to version `1.0.3`
+### Features
+* Created the component `todovue-menu`